requirements.txt
weaviate-client[agents]>=4.9.0
streamlit>=1.35.0
python-dotenv>=1.0.0
